Allow passing 3rd party install path to test script
As for the ftplibpp Makefile, use EXT_PATH to specify a single, optional
installation path for 3rd party libraries such as gtest and boost.

Note: this does not solve an issue with the test script where if one of
the CMakeLists.txt files of library or libraryTest are modified after
running the script, the changes are ignored and the previously generated
Makefile(s) get used. A more robust way to deal with these dependencies
would be via cmake find modules.